refactor: split ability workspace anchors
