+++ .sampo/changesets/1008-split-ability-workspace-anchors.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split ability workspace scaffold anchor and registry helpers into focused runtime modules without changing generated output.
